+++ ee/clickhouse/models/action.py
@@ -1,12 +1,11 @@
-import re
 from typing import Dict, List, Tuple
 
 from django.forms.models import model_to_dict
 
+from ee.clickhouse.models.property import filter_element
 from posthog.constants import AUTOCAPTURE_EVENT, TREND_FILTER_TYPE_ACTIONS
 from posthog.models import Action, Entity, Filter
 from posthog.models.action_step import ActionStep
-from posthog.models.event import Selector
 
 
 def format_action_filter(
@@ -81,59 +80,6 @@ def filter_event(step: ActionStep, prepend: str = "event", index: int = 0) -> Tu
     conditions.append("event = %({}_{})s".format(prepend, index))
 
     return conditions, params
-
-
-def _create_regex(selector: Selector) -> str:
-    regex = r""
-    for idx, tag in enumerate(selector.parts):
-        if tag.data.get("tag_name") and isinstance(tag.data["tag_name"], str):
-            if tag.data["tag_name"] == "*":
-                regex += ".+"
-            else:
-                regex += tag.data["tag_name"]
-        if tag.data.get("attr_class__contains"):
-            regex += r".*?\.{}".format(r"\..*?".join(sorted(tag.data["attr_class__contains"])))
-        if tag.ch_attributes:
-            regex += ".*?"
-            for key, value in sorted(tag.ch_attributes.items()):
-                regex += '{}="{}".*?'.format(key, value)
-        regex += r"([-_a-zA-Z0-9\.]*?)?($|;|:([^;^\s]*(;|$|\s)))"
-        if tag.direct_descendant:
-            regex += ".*"
-    return regex
-
-
-def filter_element(filters: Dict, prepend: str = "") -> Tuple[List[str], Dict]:
-    params = {}
-    conditions = []
-
-    if filters.get("selector"):
-        selector = Selector(filters["selector"], escape_slashes=False)
-        params["{}selector_regex".format(prepend)] = _create_regex(selector)
-        conditions.append("match(elements_chain, %({}selector_regex)s)".format(prepend))
-
-    if filters.get("tag_name"):
-        params["{}tag_name_regex".format(prepend)] = r"(^|;){}(\.|$|;|:)".format(filters["tag_name"])
-        conditions.append("match(elements_chain, %({}tag_name_regex)s)".format(prepend))
-
-    attributes: Dict[str, List] = {}
-
-    for key in ["href", "text"]:
-        vals = filters.get(key)
-        if filters.get(key):
-            attributes[key] = [re.escape(vals)] if isinstance(vals, str) else [re.escape(text) for text in filters[key]]
-
-    if len(attributes.keys()) > 0:
-        or_conditions = []
-        for key, value_list in attributes.items():
-            for idx, value in enumerate(value_list):
-                params["{}_{}_{}_attributes_regex".format(prepend, key, idx)] = ".*?({}).*?".format(
-                    ".*?".join(['{}="{}"'.format(key, value)])
-                )
-                or_conditions.append("match(elements_chain, %({}_{}_{}_attributes_regex)s)".format(prepend, key, idx))
-            conditions.append(" OR ".join(or_conditions))
-
-    return (conditions, params)
 
 
 def format_entity_filter(entity: Entity, prepend: str = "action", filter_by_team=True) -> Tuple[str, Dict]:
