+++ ee/clickhouse/models/property.py
@@ -1,15 +1,16 @@
+import re
 from typing import Any, Dict, List, Optional, Tuple
 
 from django.conf import settings
 from django.utils import timezone
 
 from ee.clickhouse.client import sync_execute
-from ee.clickhouse.models.action import filter_element
 from ee.clickhouse.models.cohort import format_filter_query
 from ee.clickhouse.models.util import is_int, is_json
 from ee.clickhouse.sql.events import SELECT_PROP_VALUES_SQL, SELECT_PROP_VALUES_SQL_WITH_FILTER
 from ee.clickhouse.sql.person import GET_DISTINCT_IDS_BY_PROPERTY_SQL
 from posthog.models.cohort import Cohort
+from posthog.models.event import Selector
 from posthog.models.property import Property
 from posthog.models.team import Team
 from posthog.utils import is_valid_regex, relative_date_parse
@@ -212,3 +213,56 @@ def get_property_values_for_key(key: str, team: Team, value: Optional[str] = Non
         SELECT_PROP_VALUES_SQL.format(parsed_date_from=parsed_date_from, parsed_date_to=parsed_date_to),
         {"team_id": team.pk, "key": key},
     )
+
+
+def filter_element(filters: Dict, prepend: str = "") -> Tuple[List[str], Dict]:
+    params = {}
+    conditions = []
+
+    if filters.get("selector"):
+        selector = Selector(filters["selector"], escape_slashes=False)
+        params["{}selector_regex".format(prepend)] = _create_regex(selector)
+        conditions.append("match(elements_chain, %({}selector_regex)s)".format(prepend))
+
+    if filters.get("tag_name"):
+        params["{}tag_name_regex".format(prepend)] = r"(^|;){}(\.|$|;|:)".format(filters["tag_name"])
+        conditions.append("match(elements_chain, %({}tag_name_regex)s)".format(prepend))
+
+    attributes: Dict[str, List] = {}
+
+    for key in ["href", "text"]:
+        vals = filters.get(key)
+        if filters.get(key):
+            attributes[key] = [re.escape(vals)] if isinstance(vals, str) else [re.escape(text) for text in filters[key]]
+
+    if len(attributes.keys()) > 0:
+        or_conditions = []
+        for key, value_list in attributes.items():
+            for idx, value in enumerate(value_list):
+                params["{}_{}_{}_attributes_regex".format(prepend, key, idx)] = ".*?({}).*?".format(
+                    ".*?".join(['{}="{}"'.format(key, value)])
+                )
+                or_conditions.append("match(elements_chain, %({}_{}_{}_attributes_regex)s)".format(prepend, key, idx))
+            conditions.append(" OR ".join(or_conditions))
+
+    return (conditions, params)
+
+
+def _create_regex(selector: Selector) -> str:
+    regex = r""
+    for idx, tag in enumerate(selector.parts):
+        if tag.data.get("tag_name") and isinstance(tag.data["tag_name"], str):
+            if tag.data["tag_name"] == "*":
+                regex += ".+"
+            else:
+                regex += tag.data["tag_name"]
+        if tag.data.get("attr_class__contains"):
+            regex += r".*?\.{}".format(r"\..*?".join(sorted(tag.data["attr_class__contains"])))
+        if tag.ch_attributes:
+            regex += ".*?"
+            for key, value in sorted(tag.ch_attributes.items()):
+                regex += '{}="{}".*?'.format(key, value)
+        regex += r"([-_a-zA-Z0-9\.]*?)?($|;|:([^;^\s]*(;|$|\s)))"
+        if tag.direct_descendant:
+            regex += ".*"
+    return regex
